+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
+### Added
+
+- `config set-version-check` command to enable or disable the version update check on CLI startup. Accepts `--enable` or `--disable`.
+
 ## [2.1.1] - 2026-04-08
 
 ### Changed